requirements.txt
streamlit
pandas
plotly
pymongo
python-dotenv
streamlit-authenticator
fpdf
twilio
